+++ src/main/java/uk/gov/companieshouse/api/accounts/transformer/CurrentAssetsInvestmentsTransformer.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.currentassetsinvestments.CurrentAssetsInvestmentsDataEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.currentassetsinvestments.CurrentAssetsInvestmentsEntity;
-import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.CurrentAssetsInvestments;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.currentassetsinvestments.CurrentAssetsInvestments;
 
 @Component
 public class CurrentAssetsInvestmentsTransformer implements NoteTransformer<CurrentAssetsInvestments, CurrentAssetsInvestmentsEntity> {
